# lsetCommand: merge delete and insert operations to reduce the number of memory allocations.