@antirez 2019-08-22
Hello @wubostc, thanks for your work: do you have any benchmark that shows the performance improvement? Thanks.