@wubostc 2019-08-24
@antirez ,  Performance test result:

before:
$ time python3 benchmark.py 
starting test2(10000)...
waiting for push.
waiting for lset.
the lset costs 38.83067584037781 seconds.
starting test2(10000 * 10)...
waiting for push.
waiting for lset.
the lset costs 659.2225272655487 seconds.
starting test3(10000)...
waiting for push.
waiting for lset.
the lset costs 41.85980749130249 seconds.
starting test3(10000 * 10)...
waiting for push.
waiting for lset.
the lset costs 621.0901529788971 seconds.
starting test4(10000)...
waiting for push.
waiting for lset.
the lset costs 41.686105489730835 seconds.
starting test4(10000 * 10)...
waiting for push.
waiting for lset.
the lset costs 623.4280843734741 seconds.

real	34m21.962s
user	13m56.047s
sys	3m9.855s


after:
$ time python3 benchmark.py 
starting test2(10000)...
waiting for push.
waiting for lset.
the lset costs 36.10434818267822 seconds.
starting test2(10000 * 10)...
waiting for push.
waiting for lset.
the lset costs 643.4992558956146 seconds.
starting test3(10000)...
waiting for push.
waiting for lset.
the lset costs 39.09110975265503 seconds.
starting test3(10000 * 10)...
waiting for push.
waiting for lset.
the lset costs 593.2505612373352 seconds.
starting test4(10000)...
waiting for push.
waiting for lset.
the lset costs 39.10385179519653 seconds.
starting test4(10000 * 10)...
waiting for push.
waiting for lset.
the lset costs 593.0730648040771 seconds.

real	32m59.807s
user	13m36.200s
sys	3m11.560s


so, it really helps us to improve performance on lset. XD

and, this is the test source code here.
`import redis` using [https://github.com/andymccurdy/redis-py](https://github.com/andymccurdy/redis-py) 

`
#!/usr/bin/python3

import random
import time
import redis

r = redis.Redis(host='localhost', port=6379, db=0)

key = 'key'


def test2(n):
    i = 0
    s = time.time()
    print(f'waiting for push.')
    while i < n:
        i += 1

        if (random.randint(0, 1) == 0):
            ele = str(random.randint(0, 999))
        else:
            ele = 'x' * random.randint(0, 9999) + str(random.randint(0, 999))

        if (random.randint(0, 1)):
            r.lpush(key, ele)
        else:
            r.rpush(key, ele)

    print(f'waiting for lset.')
    s = time.time()
    i = 0
    while i < n * 30:
        i += 1

        if (random.randint(0, 1) == 0):
            ele = str(random.randint(0, 999))
        else:
            ele = 'x' * random.randint(0, 9999) + str(random.randint(0, 999))

        try:
            r.lset(key, random.randint(0, n - 1), ele)
        except redis.exceptions.ResponseError:
            pass

    print(f'the lset costs {time.time() - s} seconds.')
    
    assert(r.llen(key) == n)

    r.delete(key)


def rndfortest3():
    op = random.randint(0, 6)

    if (op == 0):
        ele = 'x' * random.randint(0, 100000 - 1)
    elif (op == 1):
        ele = random.randint(0, 65536)
    elif (op == 2):
        ele = random.randint(0, 4294967296)
    elif (op == 3):
        ele = random.randint(0, 18446744073709551616)
    elif (op == 4):
        ele = -random.randint(0, 65536)
    elif (op == 5):
        ele = -random.randint(0, 4294967296)
    elif (op == 6):
        ele = -random.randint(0, 18446744073709551616)

    return str(ele)

def test3(n):
    print(f'waiting for push.')
    i = 0
    while i < n:
        i += 1
        ele = rndfortest3()

        if (random.randint(0, 1)):
            r.rpush(key, ele)
        else:
            r.lpush(key, ele)

    print(f'waiting for lset.')
    s = time.time()

    i = 0
    while i < n * 30:
        i += 1
        ele = rndfortest3()

        try:
            r.lset(key, random.randint(0, n - 1), ele)
        except redis.exceptions.ResponseError:
            pass
    
    print(f'the lset costs {time.time() - s} seconds.')

    assert(r.llen(key) == n)

    r.delete(key)


def test4(n):
    print(f'waiting for push.')
    i = 0
    while i < n:
        i += 1
        ele = i
        r.rpush(key, ele)


    print(f'waiting for lset.')
    s = time.time()
    i = 0
    while i < n * 100:
        try:
            r.lset(key, i // 100, random.randint(-i, i))
        except:
            pass

    print(f'the lset costs {time.time() - s} seconds.')

    assert(r.llen(key) == n)

    r.delete(key)


if __name__ == '__main__':

    if (r.exists(key)):
        r.delete(key)

    print('starting test2(10000)...')
    test2(10000)
    print('starting test2(10000 * 10)...')
    test2(10000 * 10)
    
    print('starting test3(10000)...')
    test3(10000)
    print('starting test3(10000 * 10)...')
    test3(10000 * 10)

    print('starting test4(10000)...')
    test3(10000)
    print('starting test4(10000 * 10)...')
    test3(10000 * 10)
`