@wubostc 2019-09-04
Hello, @antirez 
I made a mistake, that func test3 appeared four times. so i have updated the benchmark to show the performance. and i uploaded the file [benchmark.txt](https://github.com/antirez/redis/files/3572952/benchmark.txt) that it's going to be processed by python3.
Just change the extension of  `benchmark.txt` to `.py`.

and the look at the following, it seems to improve performance.
before:
```shell
time python3 benchmark.py
starting test2(10000 * 20)...
waiting for push.
waiting for lset.
the lset costs 1951.8228685855865 seconds.
starting test3(10000 * 20)...
waiting for push.
waiting for lset.
the lset costs 1656.2897427082062 seconds.
starting test4(10000 * 20)...
waiting for push.
waiting for lset.
the lset costs 2235.1750733852386 seconds.

real	98m24.050s
user	43m2.699s
sys	9m8.847s
```

after:
```shell
time python3 benchmark.py 
starting test2(10000 * 20)...
waiting for push.
waiting for lset.
the lset costs 1933.018455505371 seconds.
starting test3(10000 * 20)...
waiting for push.
waiting for lset.
the lset costs 1601.7476918697357 seconds.
starting test4(10000 * 20)...
waiting for push.
waiting for lset.
the lset costs 2195.7175319194794 seconds.

real	96m33.385s
user	42m20.643s
sys	9m5.543s
```