@sundb 2021-12-24
`ziplistReplace` has been added by https://github.com/redis/redis/commit/acb32d472da72a36dc73eca5fef1df71cffdc328,
and has been replaced by `lpReplace` it, close it.